Make `DispatchQueue.random` accessible to all tests
